Replace `dances_events` with `SubEvent` model.
A join table is insufficient to express all the information needed by the
association of dances to events. The main reason is that the (forthcoming)
`SubRound` model needs to refer to a specific dance of a multi-dance event, so
that "particular dance of an e.g. 2-dance event" needs to have its own `id`.

A secondary reason is that the ordering of each `SubEvent` is relevant, and so
needs to be tracked explicitly.