Add `SubRound` model, to which `Marks` belong.
If there are two or more dances within an event, judges' marks are not made on
the round as a whole, but rather on the individual dance. There needs to be a
way to distinguish marks for i.e. Gold American Cha Cha from those for Gold
American Rumba if the event is a Gold American Cha Cha-Rumba 2-dance.