Pull the constant definitions out into an initializer.
The canonical name for each `Section`, `Level`, and `Dance` is needed in order
to generate the seed data as well as make use of that data afterwards. Hence, it
has all been pulled out to a single initializer, which will be used for both the
DB seeder and the rest of the application. There is a circular dependency
between the DB seeder and the DB objects referred to by the constants, so the
constants pointing to the `ActiveRecord` instances are generated on-demand as
they are accessed.