Use `schema_plus` gem to create db view.
This allows adding and dropping the view to show up in  "db/schema.rb",
which lets "rake db:schema:load" work properly. "db:schema:load" only
runs the "db/schema.rb" file, and using `execute` in a migration does
not get added to "db/schema.rb".

`schema_plus` also tells the database to enforce foreign-key
constraints, so the order of seeding has to be changed so that `dances`
are added after `sections`. The `seed_fu` gem loads the files in
alphabetical order, so renaming the files guarantees the proper order.