Undo some eager association load in round display
For some reason I can't understand (probably related to the way `LEFT
OUTER JOIN` works), using the eager-loading associations caused couples
which did not receive any marks in a round to be omitted entirely.
Undoing this, and performing a separate query for each cell of the
table, resolves that problem. With enough caching, the performance
should be good enough that the extra queries don't completely kill
things.